+++ NGitLab.Mock/Clients/MergeRequestCommentClient.cs
@@ -7,10 +7,10 @@ namespace NGitLab.Mock.Clients;
 
 internal sealed class MergeRequestCommentClient : ClientBase, IMergeRequestCommentClient
 {
-    private readonly int _projectId;
-    private readonly int _mergeRequestIid;
+    private readonly long _projectId;
+    private readonly long _mergeRequestIid;
 
-    public MergeRequestCommentClient(ClientContext context, int projectId, int mergeRequestIid)
+    public MergeRequestCommentClient(ClientContext context, long projectId, long mergeRequestIid)
         : base(context)
     {
         _projectId = projectId;
